requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
annoy==1.17.3
anthropic==0.77.0
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
etelemetry==0.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
gradio==6.5.1
gradio_client==2.0.3
greenlet==3.3.1
groovy==0.1.2
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.7
idna==3.11
importlib_resources==6.5.2
instructor==1.14.5
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langchain==1.2.8
langchain-anthropic==1.3.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.1.147
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
node2vec==0.5.0
notebook==7.5.3
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==12.1.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.5.1
pooch==1.9.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
prov==2.1.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==5.3.0
pysbd==0.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
pyzmq==27.1.0
ragas==0.4.3
rdflib==7.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.14
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.0
semantic-version==2.10.0
Send2Trash==2.1.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.50.0
streamlit==1.53.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
traits==7.1.0
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
